fix(promotion) : discountRate Integer변경

+++ promotion/src/main/java/com/ohgiraffers/promotion/core/api/controller/v1/response/PromotionResponse.java
@@ -13,7 +13,7 @@ public record PromotionResponse (
         String productName,
         Integer originalPrice,
         Integer salePrice,
-        Double discountRate,
+        Integer discountRate,
         Integer totalQuantity,
         Integer soldQuantity,
         @JsonFormat(pattern = "yyyy-MM-dd'T'HH:mm:ss")
